move ZapOptionViewModel to commons
